+++ packages/parser/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @asyncapi/parser
 
+## 3.4.0
+
+### Minor Changes
+
+- e18f865: Updating jsonpath-plus dependency to mitigate CVE-2024-21534
+
+## 3.3.0
+
+### Minor Changes
+
+- bebbd39: feat: create rule `asyncapi3-channel-no-query-nor-fragment` for v3 core ruleset
+
+## 3.2.2
+
+### Patch Changes
+
+- b700a65: fix: remove forceful normalization of YAML to JSON
+
+## 3.2.1
+
+### Patch Changes
+
+- b032f3a: fix: upgrade `ajv` to version `8.17.1` to lay the foundation for fixing https://github.com/asyncapi/parser-js/issues/980 in `Studio`
+
+## 3.2.0
+
+### Minor Changes
+
+- 44331ee: Adding the parser into monorepo (Turborepo). No new features or bugfixes were introduced.
+
 ## 3.1.0
 
 ### Minor Changes